# apps/doc-engine/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Document Assembly System API - Template management and document generation
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Doc Engine API
  version: "1.0"
paths:
...
  /webhooks/signing/{provider}:
    post:
      consumes:
      - application/json
      parameters:
      - description: Provider name (e.g., documenso)
        in: path
        name: provider
        required: true
        type: string
      - description: Webhook signature (for Documenso)
        in: header
        name: X-Documenso-Secret
        type: string
      - description: Webhook signature (generic)
        in: header
        name: X-Webhook-Signature
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      summary: Handle signing provider webhook
      tags:
      - Webhooks
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"